experiment - unify gitlab/github req/get_all
