+++ all_repos/source/github.py
@@ -2,6 +2,7 @@ from typing import Dict
 from typing import NamedTuple
 
 from all_repos import github_api
+from all_repos.util import get_all
 from all_repos.util import hide_api_key_repr
 
 
@@ -20,7 +21,7 @@ class Settings(NamedTuple):
 
 
 def list_repos(settings: Settings) -> Dict[str, str]:
-    repos = github_api.get_all(
+    repos = get_all(
         f'{settings.base_url}/user/repos?per_page=100',
         headers={'Authorization': f'token {settings.api_key}'},
     )
