+++ all_repos/util.py
@@ -1,6 +1,13 @@
+import json
+import urllib.request
+from typing import Any
+from typing import Dict
 from typing import List
+from typing import NamedTuple
+from typing import Optional
 from typing import Tuple
 from typing import TYPE_CHECKING
+
 
 if TYPE_CHECKING:
     from typing import Protocol
@@ -28,3 +35,39 @@ def zsplit(bs: bytes) -> List[bytes]:
         return bs.rstrip(b'\0').split(b'\0')
     else:
         return []
+
+
+class Response(NamedTuple):
+    json: Any
+    links: Dict[str, str]
+
+
+def _parse_link(lnk: Optional[str]) -> Dict[str, str]:
+    if lnk is None:
+        return {}
+
+    ret = {}
+    parts = lnk.split(',')
+    for part in parts:
+        link, _, rel = part.partition(';')
+        link, rel = link.strip(), rel.strip()
+        assert link.startswith('<') and link.endswith('>'), link
+        assert rel.startswith('rel="') and rel.endswith('"'), rel
+        link, rel = link[1:-1], rel[len('rel="'):-1]
+        ret[rel] = link
+    return ret
+
+
+def req(url: str, **kwargs: Any) -> Response:
+    resp = urllib.request.urlopen(urllib.request.Request(url, **kwargs))
+    return Response(json.load(resp), _parse_link(resp.headers['link']))
+
+
+def get_all(url: str, **kwargs: Any) -> List[Dict[str, Any]]:
+    ret: List[Dict[str, Any]] = []
+    resp = req(url, **kwargs)
+    ret.extend(resp.json)
+    while 'next' in resp.links:
+        resp = req(resp.links['next'], **kwargs)
+        ret.extend(resp.json)
+    return ret
