--- tests/gitlab_api_test.py
@@ -1,22 +1,0 @@
-from all_repos.gitlab_api import get_all
-from testing.mock_http import FakeResponse
-from testing.mock_http import urlopen_side_effect
-
-
-def test_get_all(mock_urlopen):
-    mock_urlopen.side_effect = urlopen_side_effect({
-        'https://example.com/api': FakeResponse(
-            b'["page1_1", "page1_2"]',
-            next_link='https://example.com/api?page=2',
-        ),
-        'https://example.com/api?page=2': FakeResponse(
-            b'["page2_1", "page2_2"]',
-            next_link='https://example.com/api?page=3',
-        ),
-        'https://example.com/api?page=3': FakeResponse(
-            b'["page3_1"]',
-        ),
-    })
-
-    ret = get_all('https://example.com/api')
-    assert ret == ['page1_1', 'page1_2', 'page2_1', 'page2_2', 'page3_1']
